refactor: Move GTFSNnn into gtfs graph builder
