+++ src/main/java/org/opentripplanner/gtfs/graphbuilder/GtfsFeedParametersBuilder.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.standalone.config.feed;
+package org.opentripplanner.gtfs.graphbuilder;
 
 import java.net.URI;
